requirements.txt
Django==3.0.2
djangorestframework==3.12.1
python_dateutil==2.8.1
